fix mouse for chrome (for now)
Seems scroll delta is inconsistent between browsers, so for my chromium
installation it was simply not working at all due to snapping. Made it
so every event is a cursor size change for now. probably bad for smooth
mouse wheels, but for a complete fix we would have to keep track of
pixels scrolled and probably add a mouse wheel sensitivity setting
somewhere.

Signed-off-by: Victor Seiji Hariki <victorseijih@gmail.com>